fix typescript problems
