+++ src/gdrive/tools/index.ts
@@ -1,0 +1,23 @@
+import { schema as gdriveSearchSchema, search } from './gdrive_search.js';
+import { schema as gdriveReadFileSchema, readFile } from './gdrive_read_file.js';
+import { schema as gsheetsUpdateCellSchema, updateCell } from './gsheets_update_cell.js';
+import { Tool, GDriveSearchInput, GDriveReadFileInput, GSheetsUpdateCellInput } from './types.js';
+
+export const tools: [
+  Tool<GDriveSearchInput>,
+  Tool<GDriveReadFileInput>,
+  Tool<GSheetsUpdateCellInput>
+] = [
+  {
+    ...gdriveSearchSchema,
+    handler: search,
+  },
+  {
+    ...gdriveReadFileSchema,
+    handler: readFile,
+  },
+  {
+    ...gsheetsUpdateCellSchema,
+    handler: updateCell,
+  },
+];
